requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
python-dotenv
langchain-openai
langchain-classic
langchain_mcp_adapters
